Plumb v2 shim through tests.
This change updates test/root/crictl_test.go and related utilities, so
that the containerd tests can be run on any version (and in cases where
it applies, they can be run on both v1 and v2 as parameterized tests).

This change also makes the containerd installation dynamic, so that
multiple versions can be tested, and drops the static installer for the
VM image itself.