Add NewJob to tcpip.Clock
Changes the API of tcpip.Clock to also provide a method for scheduling and
rescheduling work after a specified duration. This change also implements the
NewJob method for existing implementations of tcpip.Clock.

This is the groundwork required to mock time within tests. All references to
CancellableTimer has been replaced with the tcpip.Job interface, allowing for
custom implementations of scheduling work.

This is a BREAKING CHANGE for clients that implement their own tcpip.Clock or
use tcpip.CancellableTimer. Migration plan:
 1. Implement tcpip.Job
 2. Add the NewJob method to tcpip.Clock
 3. Replace references of tcpip.CancellableTimer with tcpip.Job
 4. Replace calls to tcpip.CancellableTimer#StopLocked with tcpip.Job#Cancel
 5. Replace calls to tcpip.CancellableTimer#Reset with tcpip.Job#Schedule
 6. Replace calls to tcpip.NewCancellableTimer with tcpip.Clock#NewJob

PiperOrigin-RevId: 316767762